% Generated by build script
\name{meteo_geocode}
\alias{meteo_geocode}
\title{Geocode a location name to coordinates}
\usage{
meteo_geocode()
}
\arguments{
\item{name}{Place name (e.g., "New York", "Tokyo", "London")}
\item{count}{Number of results (default 5)}
}
\value{tibble: name, latitude, longitude, country, timezone, elevation, population}
\description{Geocode a location name to coordinates}
