% Generated by build script
\name{meteo_historical}
\alias{meteo_historical}
\title{Get historical weather archive data}
\usage{
meteo_historical()
}
\arguments{
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{start_date}{Start date (required, YYYY-MM-DD)}
\item{end_date}{End date (required, YYYY-MM-DD)}
\item{daily}{Variables (same options as meteo_daily)}
}
\value{tibble with date and requested variables}
\description{Get historical weather archive data}
