% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{meteo_historical}
\alias{meteo_historical}
\title{Get historical weather archive data}
\usage{
meteo_historical(
  latitude,
  longitude,
  start_date,
  end_date,
  daily = "temperature_2m_max,temperature_2m_min,precipitation_sum"
)
}
\arguments{
\item{latitude}{Latitude}

\item{longitude}{Longitude}

\item{start_date}{Start date (required, YYYY-MM-DD)}

\item{end_date}{End date (required, YYYY-MM-DD)}

\item{daily}{Variables (same options as meteo_daily)}
}
\value{
tibble with date and requested variables
}
\description{
Access historical weather data from 1940 to present.
}
